Moved most files into visualisation subdirectory.
Deleted unused quanternion.h (its .cpp was missing)

Deleted resources.h, garbage generated by vb at project creation.
