+++ sdl_exp/main.cpp
@@ -1,4 +1,4 @@
-#include "Visualisation.h"
+#include "visualisation/Visualisation.h"
 
 int main()
 {
