+++ sdl_exp/visualisation/Entity.cpp
