+++ sdl_exp/visualisation/Visualisation.cpp
@@ -1,7 +1,8 @@
 #include "Visualisation.h"
 
 #include <sstream>
-#include <glm\gtc\type_ptr.hpp>
+#include <glm/gtc/type_ptr.hpp>
+#include <glm/gtc/matrix_transform.inl>
 
 #include "GLcheck.h"
