+++ sdl_exp/visualisation/VisualisationScene.cpp
